requirements.txt
python-telegram-bot==20.0
pymongo==4.3.3
python-dotenv==1.0.0
cloudinary
Pillow==9.5.0
psycopg2-binary==2.9.5
apscheduler==3.9.1
python-dateutil==2.8.2